+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.8.3 - Unreleased
 
 - Added `code_sum` function for efficient combined Data-Code and Instance-Code generation
+- Changed `code_iscc` to compute Data & Instance Code in one go
 
 ## 0.8.2 - 2025-05-02